restructuring of the dirtree
