+++ src/include/def.h
@@ -21,4 +21,18 @@ typedef unsigned char       u8;
 #define __packed        __attribute__((packed))
 #define force_inline    __attribute__((always_inline)) inline
 
+#define IN_BOUNDS(LEFT, A, RIGHT)   ((LEFT) < (A) && (A) < (RIGHT))
+#define IN_BOUNDS_EQ(LEFT, A, RIGHT) ((LEFT) <= (A) && (A) <= (RIGHT))
+
+#define ABS(A)  ((A) >= 0 ? (A) : -(A))
+#define SGN(A)  ((A) >= 0 ? 1 : -1)
+
+#define ARR_SIZE(arr)   ((sizeof(arr)) / (sizeof(arr[0])))
+
+#define LOW_OFFSET(off)     ((off) & 0xff)
+#define HIGH_OFFSET(off)    (((off) >> 8) & 0xff)
+#define low_16(addr)    (uint16_t)(((addr) & 0xFFFF));
+#define high_16(addr)   (uint16_t)(((addr) >> 16) & 0xFFFF);
+
+
 #endif
