+++ src/libs/array.h
