+++ src/libs/string.h
